+++ @types/yggdrassil/service/clients/index.d.ts
@@ -1,0 +1,2 @@
+export * from "./rest-client";
+//# sourceMappingURL=index.d.ts.map