+++ @types/yggdrassil/service/routes/belo-boom.d.ts
@@ -1,0 +1,4 @@
+import { FastifyPluginAsync } from "fastify";
+export declare const fastifyBeloBoom: FastifyPluginAsync<Record<never, never>, import("fastify").RawServerDefault, import("fastify").FastifyTypeProviderDefault, import("fastify").FastifyBaseLogger>;
+export default fastifyBeloBoom;
+//# sourceMappingURL=belo-boom.d.ts.map