+++ @types/yggdrassil/utils/data/mcc.d.ts
@@ -1,0 +1,12 @@
+export interface MccData {
+    mcc: string;
+    edited_description: string;
+    combined_description: string;
+    usda_description: string;
+    irs_description: string;
+    irs_reportable: string;
+    is_blacklisted?: boolean;
+    id: number;
+}
+export declare const mccData: MccData[];
+//# sourceMappingURL=mcc.d.ts.map