+++ @types/yggdrassil/utils/database.d.ts
@@ -1,0 +1,3 @@
+import { Knex } from "knex";
+export declare function commonColumns(knex: Knex, table: Knex.CreateTableBuilder): void;
+//# sourceMappingURL=database.d.ts.map