+++ @types/yggdrassil/utils/lambda.d.ts
@@ -1,0 +1,3 @@
+import { InvocationRequest } from "aws-sdk/clients/lambda";
+export declare const invoke: (parameters: InvocationRequest) => Promise<unknown>;
+//# sourceMappingURL=lambda.d.ts.map