+++ @types/yggdrassil/utils/mongoose.d.ts
@@ -1,0 +1,2 @@
+export declare const connectToMongo: (url: string) => Promise<void>;
+//# sourceMappingURL=mongoose.d.ts.map