+++ @types/yggdrassil/utils/opentelemetry.d.ts
@@ -1,0 +1,6 @@
+export declare class Opentelemetry {
+    private sdk?;
+    constructor(resourceName: string, traceExporterUrl: string);
+    start(): Promise<void | undefined>;
+}
+//# sourceMappingURL=opentelemetry.d.ts.map