+++ @types/yggdrassil/utils/qr/checksum.d.ts
@@ -1,0 +1,3 @@
+export declare function computeCRC(value: string): string;
+export declare function validateChecksum(emvString: string): boolean;
+//# sourceMappingURL=checksum.d.ts.map