+++ @types/yggdrassil/utils/qr/models/schema-additional-fields.d.ts
@@ -1,0 +1,48 @@
+declare const _default: {
+    "01": {
+        name: string;
+        presence: string;
+    };
+    "02": {
+        name: string;
+        presence: string;
+    };
+    "03": {
+        name: string;
+        presence: string;
+    };
+    "04": {
+        name: string;
+        presence: string;
+    };
+    "05": {
+        name: string;
+        presence: string;
+    };
+    "06": {
+        name: string;
+        presence: string;
+    };
+    "07": {
+        name: string;
+        presence: string;
+    };
+    "08": {
+        name: string;
+        presence: string;
+    };
+    "09": {
+        name: string;
+        presence: string;
+    };
+    "10-49": {
+        name: string;
+        presence: string;
+    };
+    "50-99": {
+        name: string;
+        presence: string;
+    };
+};
+export default _default;
+//# sourceMappingURL=schema-additional-fields.d.ts.map