+++ @types/yggdrassil/utils/qr/models/schema-sub-data.d.ts
@@ -1,0 +1,20 @@
+declare const _default: {
+    "00": {
+        name: string;
+        presence: string;
+    };
+    "01": {
+        name: string;
+        presence: string;
+    };
+    "02": {
+        name: string;
+        presence: string;
+    };
+    "03": {
+        name: string;
+        presence: string;
+    };
+};
+export default _default;
+//# sourceMappingURL=schema-sub-data.d.ts.map