+++ @types/yggdrassil/utils/time.d.ts
@@ -1,0 +1,2 @@
+export declare const getBinDateRanges: (startDate: string, endDate: string, rangeLengthInDays: string) => [string, string][];
+//# sourceMappingURL=time.d.ts.map